Fix average runtime calculation to use int for numerator

+++ Python dla szkoł/11.Typy danych i konwersja/4.py
@@ -1,2 +1,2 @@
 def avgRuntime( x ):
-    return float(float(x[:3]) / int(x[4:]))
+    return float(int(x[:3]) / int(x[4:]))
